update to support less than 1 percent random

+++ Objects/Global/Random/Interface/IRandom.cs
@@ -11,6 +11,17 @@ namespace Objects.Global.Random.Interface
 
         int Next(int minValue, int maxValue);
 
+        ///// <summary>
+        ///// Takes a percentage and rolls a dice to see
+        ///// if the roll is less than/equal to the percent.  
+        ///// If so it returns true.
+        ///// IE 20 returns true 20% of the time and
+        ///// 100 returns true 100% of the time.
+        ///// </summary>
+        ///// <param name="percentSuccessful"></param>
+        ///// <returns></returns>
+        //bool PercentDiceRoll(int percentSuccessful);
+
         /// <summary>
         /// Takes a percentage and rolls a dice to see
         /// if the roll is less than/equal to the percent.  
@@ -20,6 +31,6 @@ namespace Objects.Global.Random.Interface
         /// </summary>
         /// <param name="percentSuccessful"></param>
         /// <returns></returns>
-        bool PercentDiceRoll(int percentSuccessful);
+        bool PercentDiceRoll(double percentSuccessful);
     }
 }
